requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Database
motor==3.5.1
pymongo==4.8.0

# Data handling (pin to stable versions that have prebuilt wheels for Python 3.11/3.12)
pandas==2.2.2
numpy==1.26.4

# Visualization
plotly==5.23.0
kaleido==0.2.1   # required for exporting charts to images

# Environment variables
python-dotenv==1.0.1

# OpenAI SDK
openai==1.51.0

# Validation
pydantic==2.8.2

# Logging helpers
rich==13.7.1
